Fold the ant snapshot save/load into ant_colony.h
