+++ src/Qubic.vcxproj
@@ -68,7 +68,6 @@
     <ClInclude Include="logging\net_msg_impl.h" />
     <ClInclude Include="mining\ant_colony\ant_colony.h" />
     <ClInclude Include="mining\ant_colony\ant_colony_bpp9000.h" />
-    <ClInclude Include="mining\ant_colony\ant_colony_snapshot.h" />
     <ClInclude Include="mining\ant_colony\ant_pending_solutions.h" />
     <ClInclude Include="mining\custom_qubic_mining_storage.h" />
     <ClInclude Include="mining\mining.h" />
