+++ src/Qubic.vcxproj.filters
@@ -197,9 +197,6 @@
     <ClInclude Include="mining\ant_colony\ant_colony_bpp9000.h">
       <Filter>mining\ant_colony</Filter>
     </ClInclude>
-    <ClInclude Include="mining\ant_colony\ant_colony_snapshot.h">
-      <Filter>mining\ant_colony</Filter>
-    </ClInclude>
     <ClInclude Include="mining\trit_pack.h">
       <Filter>mining</Filter>
     </ClInclude>
